style:make dashboard chart mobile friendly
