+++ frontend/src/app/components/OverViewCard.tsx
@@ -19,14 +19,16 @@ const OverViewCard: React.FC<IOverViewCard> = ({
     return (
         <Card
             className={cn(
-                'hover-transition-primary flex h-full w-full flex-col justify-between gap-y-0 p-4 pb-6',
+                'hover-transition-primary flex h-full w-full flex-col justify-between p-4 pb-6 min-gap-y-2 gap-y-2',
                 className
             )}
         >
             <div className="flex flex-col gap-y-2">
                 <div className="h4">{title}</div>
                 <div className="card-h1 pl-[2px]">{value}</div>
-            {children}
+            </div>
+            <div>
+                {children}
             </div>
         </Card>
     );
